+++ tests/commands/test_schema_report.py
@@ -4,16 +4,9 @@ from unittest.mock import patch
 
 from ocdskit.cli.__main__ import main
 
+# Taken from test_set_closed_codelist_enums.py
 stdin = b'''{
   "properties": {
-    "open": {
-      "type": [
-        "string",
-        "null"
-      ],
-      "codelist": "a.csv",
-      "openCodelist": true
-    },
     "closedStringNull": {
       "type": [
         "string",
@@ -47,11 +40,13 @@ stdin = b'''{
         "foo"
       ]
     },
-    "definition": {
+    "open": {
       "type": [
         "string",
         "null"
-      ]
+      ],
+      "codelist": "a.csv",
+      "openCodelist": true
     }
   }
 }
@@ -60,14 +55,39 @@ stdin = b'''{
 
 def test_command(monkeypatch):
     with patch('sys.stdin', TextIOWrapper(BytesIO(stdin))), patch('sys.stdout', new_callable=StringIO) as actual:
-        monkeypatch.setattr(sys, 'argv', ['ocdskit', 'schema-report'])
+        monkeypatch.setattr(sys, 'argv', ['ocdskit', 'schema-report', '--min-occurrences', '2'])
         main()
 
-    assert actual.getvalue() == '''openCodelist: False
-a.csv
-b.csv
-c.csv
-d.csv
+    assert actual.getvalue() == '''codelist,openness
+a.csv,closed/open
+b.csv,closed
+c.csv,closed
+d.csv,closed
 
- 3: {'type': ['string', 'null']}
+ 2: {'type': ['string', 'null']}
+ 2: {'type': 'string'}
 '''
+
+
+def test_command_no_codelists(monkeypatch):
+    with patch('sys.stdin', TextIOWrapper(BytesIO(stdin))), patch('sys.stdout', new_callable=StringIO) as actual:
+        monkeypatch.setattr(sys, 'argv', ['ocdskit', 'schema-report', '--min-occurrences', '2', '--no-codelists'])
+        main()
+
+    assert 'codelist' not in actual.getvalue()
+
+
+def test_command_no_definitions(monkeypatch):
+    with patch('sys.stdin', TextIOWrapper(BytesIO(stdin))), patch('sys.stdout', new_callable=StringIO) as actual:
+        monkeypatch.setattr(sys, 'argv', ['ocdskit', 'schema-report', '--min-occurrences', '2', '--no-definitions'])
+        main()
+
+    assert ':' not in actual.getvalue()
+
+
+def test_command_min_occurrences(monkeypatch):
+    with patch('sys.stdin', TextIOWrapper(BytesIO(stdin))), patch('sys.stdout', new_callable=StringIO) as actual:
+        monkeypatch.setattr(sys, 'argv', ['ocdskit', 'schema-report', '--min-occurrences', '1', '--no-codelists'])
+        main()
+
+    assert '1:' in actual.getvalue()